// src/enums.d.ts
export declare enum Geography {
    Nordjylland = "Nordjylland",
    Midtjylland = "Midtjylland",
    Hovedstaden = "Hovedstaden",
    Sjælland = "Sj\u00E6lland",
    Syddanmark = "Syddanmark",
    Ukendt = "Ukendt",
    Uoplyst = "Uoplyst"
}
export declare enum Institution {
    "Aalborg Universitet" = "Aalborg Universitet",
    "Aarhus Maskinmesterskole" = "Aarhus Maskinmesterskole",
    "Aarhus Universitet" = "Aarhus Universitet",
    "Arkitektskolen Aarhus" = "Arkitektskolen Aarhus",
    "Copenhagen Business School - Handelshøjskolen" = "Copenhagen Business School - Handelsh\u00F8jskolen",
    "Dalum Landbrugsskole" = "Dalum Landbrugsskole",
    "Danmarks Medie- og Journalisthøjskole" = "Danmarks Medie- og Journalisth\u00F8jskole",
    "Danmarks Tekniske Universitet" = "Danmarks Tekniske Universitet",
    "Den Danske Filmskole" = "Den Danske Filmskole",
    "Den Danske Scenekunstskole" = "Den Danske Scenekunstskole",
    "Den Frie Lærerskole" = "Den Frie L\u00E6rerskole",
    "Designskolen Kolding" = "Designskolen Kolding",
    "Det Fynske Kunstakademi" = "Det Fynske Kunstakademi",
    "Det Jyske Kunstakademi" = "Det Jyske Kunstakademi",
    "Det Jyske Musikkonservatorium" = "Det Jyske Musikkonservatorium",
    "Det Kgl. Danske Musikkonservatorium" = "Det Kgl. Danske Musikkonservatorium",
    "Det Kongelige Akademi - Arkitektur, Design, Konservering" = "Det Kongelige Akademi - Arkitektur, Design, Konservering",
    "Det Kongelige Danske Kunstakademi - Billedkunstskolerne" = "Det Kongelige Danske Kunstakademi - Billedkunstskolerne",
    "Diakonissestiftelsen" = "Diakonissestiftelsen",
    "EUC Nordvest" = "EUC Nordvest",
    "Erhvervsakademi Aarhus" = "Erhvervsakademi Aarhus",
    "Erhvervsakademi Dania" = "Erhvervsakademi Dania",
    "Erhvervsakademi MidtVest" = "Erhvervsakademi MidtVest",
    "Erhvervsakademi SydVest" = "Erhvervsakademi SydVest",
    "Erhvervsakademiet Copenhagen Business Academy" = "Erhvervsakademiet Copenhagen Business Academy",
    "Forfatterskolen" = "Forfatterskolen",
    "Fredericia Maskinmesterskole" = "Fredericia Maskinmesterskole",
    "IBA Erhvervsakademi Kolding" = "IBA Erhvervsakademi Kolding",
    "IT-Universitetet i København" = "IT-Universitetet i K\u00F8benhavn",
    "Kriminalforsorgens Uddannelsescenter" = "Kriminalforsorgens Uddannelsescenter",
    "Københavns Erhvervsakademi (KEA)" = "K\u00F8benhavns Erhvervsakademi (KEA)",
    "Københavns Professionshøjskole" = "K\u00F8benhavns Professionsh\u00F8jskole",
    "Københavns Universitet" = "K\u00F8benhavns Universitet",
    "MARTEC - Maritime and Polytechnic College" = "MARTEC - Maritime and Polytechnic College",
    "Marstal Navigationsskole" = "Marstal Navigationsskole",
    "Maskinmesterskolen København" = "Maskinmesterskolen K\u00F8benhavn",
    "Møgelkær Fængsel" = "M\u00F8gelk\u00E6r F\u00E6ngsel",
    "Nyborg Søfartsskole" = "Nyborg S\u00F8fartsskole",
    "Politiskolen" = "Politiskolen",
    "Politiskolen, Uddannelsescenter Vest (UCV)" = "Politiskolen, Uddannelsescenter Vest (UCV)",
    "Professionshøjskolen Absalon" = "Professionsh\u00F8jskolen Absalon",
    "Professionshøjskolen UC Syddanmark" = "Professionsh\u00F8jskolen UC Syddanmark",
    "Professionshøjskolen University College Nordjylland" = "Professionsh\u00F8jskolen University College Nordjylland",
    "Professionshøjskolen VIA University College" = "Professionsh\u00F8jskolen VIA University College",
    "Roskilde Universitet" = "Roskilde Universitet",
    "Rytmisk Musikkonservatorium" = "Rytmisk Musikkonservatorium",
    "Skoleskibet Georg Stage" = "Skoleskibet Georg Stage",
    "Svendborg International Maritime Academy, SIMAC" = "Svendborg International Maritime Academy, SIMAC",
    "Svendborg Søfartsskole" = "Svendborg S\u00F8fartsskole",
    "Syddansk Musikkonservatorium" = "Syddansk Musikkonservatorium",
    "Syddansk Universitet" = "Syddansk Universitet",
    "TietgenSkolen" = "TietgenSkolen",
    "UCL Erhvervsakademi og Professionshøjskole" = "UCL Erhvervsakademi og Professionsh\u00F8jskole",
    "Uddannelsen på landsplan" = "Uddannelsen p\u00E5 landsplan",
    "Vildtforvaltningsskolen, Kalø" = "Vildtforvaltningsskolen, Kal\u00F8",
    "Zealand Sjællands Erhvervsakademi" = "Zealand Sj\u00E6llands Erhvervsakademi",
    "?" = "?"
}
export declare enum DegreeType {
    "Akademisk overbygningsuddannelse" = "Akademisk overbygningsuddannelse",
    "Bacheloruddannelse" = "Bacheloruddannelse",
    "Erhvervsakademiuddannelse" = "Erhvervsakademiuddannelse",
    "Kandidatuddannelse" = "Kandidatuddannelse",
    "Kunstnerisk uddannelse" = "Kunstnerisk uddannelse",
    "Politi og forsvar" = "Politi og forsvar",
    "Professionsbacheloruddannelse" = "Professionsbacheloruddannelse",
    "Andre uddannelser" = "Andre uddannelser"
}
export declare enum County {
    Ballerup = "Ballerup",
    Bornholm = "Bornholm",
    Brøndby = "Br\u00F8ndby",
    Esbjerg = "Esbjerg",
    "Faaborg-Midtfyn" = "Faaborg-Midtfyn",
    Fredericia = "Fredericia",
    Frederiksberg = "Frederiksberg",
    Frederikshavn = "Frederikshavn",
    Guldborgsund = "Guldborgsund",
    Haderslev = "Haderslev",
    Hedensted = "Hedensted",
    Herning = "Herning",
    Hillerød = "Hiller\u00F8d",
    Hjørring = "Hj\u00F8rring",
    Holbæk = "Holb\u00E6k",
    Holstebro = "Holstebro",
    Horsens = "Horsens",
    "Ikast-Brande" = "Ikast-Brande",
    Kalundborg = "Kalundborg",
    Kolding = "Kolding",
    København = "K\u00F8benhavn",
    Køge = "K\u00F8ge",
    Lemvig = "Lemvig",
    "Lyngby-Taarbæk" = "Lyngby-Taarb\u00E6k",
    Mariagerfjord = "Mariagerfjord",
    Norddjurs = "Norddjurs",
    Næstved = "N\u00E6stved",
    Odense = "Odense",
    Randers = "Randers",
    Roskilde = "Roskilde",
    Rudersdal = "Rudersdal",
    Silkeborg = "Silkeborg",
    Skive = "Skive",
    Slagelse = "Slagelse",
    Svendborg = "Svendborg",
    Syddjurs = "Syddjurs",
    Sønderborg = "S\u00F8nderborg",
    Thisted = "Thisted",
    Tønder = "T\u00F8nder",
    Vejle = "Vejle",
    Viborg = "Viborg",
    Vordingborg = "Vordingborg",
    Ærø = "\u00C6r\u00F8",
    Aabenraa = "Aabenraa",
    Aalborg = "Aalborg",
    Aarhus = "Aarhus",
    Ukendt = "Ukendt",
    Uoplyst = "Uoplyst"
}
export declare enum SubjectTitle {
    Matematik = "Matematik",
    Dansk = "Dansk",
    Fysik = "Fysik",
    Kemi = "Kemi",
    Biologi = "Biologi",
    Historie = "Historie",
    Idræt = "Idr\u00E6t"
}